### Description of Change

The routine of merging latest changes to net8.0 try to bring them to rc1

+++ src/Core/src/Platform/Android/BorderDrawable.cs
@@ -17,7 +17,7 @@ namespace Microsoft.Maui.Platform
 	public class BorderDrawable : PaintDrawable
 	{
 		readonly AContext? _context;
-		readonly double _density;
+		readonly float _density;
 
 		bool _invalidatePath;
 
@@ -292,6 +292,7 @@ namespace Microsoft.Maui.Platform
 			{
 				_invalidatePath = false;
 
+				// TODO: Should this do the same thing as MauiDrawable.Android.cs? I suspect it should.
 				var path = GetPath(_width, _height, _cornerRadius, _strokeThickness);
 				var clipPath = path?.AsAndroidPath();
